feat: reorganize types and export in data package
