+++ apps/web/public/robots.txt
@@ -2,4 +2,4 @@ User-agent: *
 Disallow: /questionnaire/
 Disallow: /ajouter-une-aide-entreprises
 
-Sitemap: undefined/sitemap.xml
+Sitemap: http://localhost:4242/sitemap.xml
