+++ libs/data/src/index.ts
@@ -1,6 +1,6 @@
 export * from './program/program'
-export * from './program/types'
+export * from './program/types/export'
 export * from './type/publicodesTypes'
 export * from './generated/program'
-export * from './theme/themes'
-export * from './project/types'
+export * from './theme/types/export'
+export * from './project/types/export'
