+++ libs/data/src/program/types/export.ts
@@ -1,0 +1,21 @@
+export enum ProgramAidType {
+  acc = 'accompagnement',
+  train = 'formation',
+  fund = 'financement',
+  loan = 'prêt',
+  tax = 'avantage fiscal'
+}
+
+export enum ProgramOperatorType {
+  ADEME = 'ADEME',
+  ASP = 'ASP',
+  BPI = 'Bpifrance',
+  CCI = 'CCI',
+  CMA = 'CMA',
+  DDFIP = 'DDFIP',
+  LAPOSTE = 'La Poste',
+  MTES = 'Ministère de la Transition Écologique et Solidaire',
+  ORACE = 'ORACE',
+  BREIZHFAB = 'Breizh Fab',
+  ECOCO2 = 'EcoCO2'
+}
