+++ libs/data/src/project/types/domain.ts
@@ -1,0 +1,8 @@
+import { Project } from './export'
+
+export interface RawProject extends Omit<Project, 'themes' | 'mainTheme' | 'linkedProjects' | 'programs'> {
+  themes: string[]
+  mainTheme: string
+  linkedProjects: number[]
+  programs: string[]
+}
