+++ libs/data/src/project/types/export.ts
@@ -1,5 +1,5 @@
-import { ThemeId } from '../theme/themes'
-import { projects } from '../../static'
+import { ThemeId } from '../../theme/types/export'
+import { projects } from '../../../static'
 
 export interface Project {
   id: number
@@ -18,10 +18,3 @@ export interface Project {
 }
 
 export type ProjectId = (typeof projects)[number]['id']
-
-export interface RawProject extends Omit<Project, 'themes' | 'mainTheme' | 'linkedProjects' | 'programs'> {
-  themes: string[]
-  mainTheme: string
-  linkedProjects: number[]
-  programs: string[]
-}
