+++ libs/data/src/theme/types/domain.ts
@@ -1,0 +1,7 @@
+import { ThemeId } from './export'
+
+export interface Theme {
+  id: number
+  Nom: string
+  'Nom (Tech)': ThemeId
+}
