+++ libs/data/static/index.ts
@@ -1,4 +1,4 @@
-import { Project } from '../src/project/types'
+import { Project } from '../src/project/types/export'
 import { default as projectsJson } from './projects.json'
 
 export const projects = projectsJson as unknown as Project[]
